All hexagons are placed to the map. LOOK AT THE DETAILS OF COMMIT
Issue1: All hexagons are placed to the screen but there are some little spaces between the hexagons. Probably after setting the vertices and edges in java file (instead of fxml) they will be placed at top of the hexagons so that the spaces will disappear.
Issue: When GameMenu opens, it stays behind of the hexagons so it also has to be fixed.